fix: move StreamClickableNodes behind darwin build tag
